feat: Support Class, Function 16

+++ src/test/java/com/caoccao/javet/swc4j/compiler/ast/clazz/clazz/TestCompileAstClassFields.java
@@ -89,8 +89,8 @@ public class TestCompileAstClassFields extends BaseTestCompileSuite {
                         "string", "Hello"
                 ),
                 Map.of(
-                        "int", (int) instanceRunner.invoke("getInt"),
-                        "double", (double) instanceRunner.invoke("getDouble"),
+                        "int", instanceRunner.invoke("getInt"),
+                        "double", instanceRunner.invoke("getDouble"),
                         "bool", (boolean) instanceRunner.invoke("getBool"),
                         "string", (String) instanceRunner.invoke("getString")
                 )
